workspace/prueba4/requirements.txt
fastapi==0.110.*
uvicorn[standard]==0.27.*
python-multipart==0.0.9
SQLAlchemy==2.0.*
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.*